Fix cast of uint32 to void * so it should work on both 32 and 64 bit archs.

+++ src/control/midi/portmidi/pm_mac/pmmacosxcm.c
@@ -30,6 +30,7 @@
 #include "pmmacosxcm.h"
 
 #include <stdio.h>
+#include <stdint.h>	// for uintptr_t
 #include <string.h>
 
 #include <CoreServices/CoreServices.h>
@@ -666,7 +667,7 @@ PmError pm_macosxcm_init(void)
         
         /* Register this device with PortMidi */
         pm_add_device("CoreMIDI", cm_get_full_endpoint_name(endpoint),
-                      TRUE, (void*)endpoint, &pm_macosx_in_dictionary);
+                      TRUE, (void*)((uintptr_t)endpoint), &pm_macosx_in_dictionary);
     }
 
     /* Iterate over the MIDI output devices */
@@ -682,7 +683,7 @@ PmError pm_macosxcm_init(void)
 
         /* Register this device with PortMidi */
         pm_add_device("CoreMIDI", cm_get_full_endpoint_name(endpoint),
-                      FALSE, (void*)endpoint, &pm_macosx_out_dictionary);
+                      FALSE, (void*)((uintptr_t)endpoint), &pm_macosx_out_dictionary);
     }
     return pmNoError;
